requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
pydantic-settings==2.0.3
openai==1.3.0
replicate==0.18.0
httpx==0.25.2
boto3==1.34.0
